Create interface for spring service
